+++ lib/IE/component/IComponent/IScalable/BasicInteractiveData/BaseButton/ButtonPanel/ButtonPanel.cpp
@@ -63,7 +63,7 @@ namespace ie {
 		return new ButtonPanel{*this};
 	}
 	
-	void ButtonPanel::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void ButtonPanel::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		BaseButton::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		panel->draw_debug(render_target, 0, indent_addition, hue, hue_offset);
 	}
