+++ lib/IE/component/IComponent/IScalable/Box/BoxConstBorder/BoxConstBorder.cpp
@@ -101,7 +101,7 @@ namespace ie {
 		return new BoxConstBorder{*this};
 	}
 	
-	void BoxConstBorder::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void BoxConstBorder::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		IComponent::draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 		const_object->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
 		second_object->draw_debug(render_target, indent + indent_addition, indent_addition, hue + hue_offset, hue_offset);
