+++ lib/IE/component/IComponent/IScalable/Box/BoxConstBorder/BoxConstBorder.hpp
@@ -44,7 +44,7 @@ namespace ie {
 		
 		BoxConstBorder* copy() override;
 		
-		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) override;
+		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) override;
 	
 	protected:
 		BoxPtr<IScalable> const_object;
