+++ lib/IE/component/IComponent/IScalable/Box/BoxMovableBorder/BoxMovableBorder.hpp
@@ -80,7 +80,7 @@ namespace ie {
 		
 		BoxMovableBorder* copy() override;
 		
-		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) override;
+		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) override;
 	
 	protected:
 		BasicInteractiveData<BoxMovableBorder&> interactive;
