+++ lib/IE/component/IComponent/IScalable/IUninteractive/Caption/Caption.hpp
@@ -100,7 +100,7 @@ namespace ie {
 		
 		Caption* copy() override;
 		
-		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) override;
+		void draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) override;
 	
 	protected:
 		static sf::Color default_color;
