+++ lib/IE/component/IComponent/IScalable/IUninteractive/OnlyDrawable/Empty/Empty.hpp
@@ -26,7 +26,7 @@ namespace ie {
 		
 		Empty* copy() override;
 		
-		void draw_debug(sf::RenderTarget&, int, int, uint, uint) override;
+		void draw_debug(sf::RenderTarget&, int, int, size_t, size_t) override;
 	
 	protected:
 		LayoutData& layout_get_data() override;
