+++ lib/IE/component/IComponent/IScalable/Interface/Interface.cpp
@@ -194,7 +194,7 @@ namespace ie {
 		return interface;
 	}
 	
-	void Interface::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, uint hue, uint hue_offset) {
+	void Interface::draw_debug(sf::RenderTarget& render_target, int indent, int indent_addition, size_t hue, size_t hue_offset) {
 		object->draw_debug(render_target, indent, indent_addition, hue, hue_offset);
 	}
