+++ lib/IE/component/IComponent/Text/BaseCharacter/ObjectCharacter/ObjectCharacter.hpp
@@ -36,7 +36,7 @@ namespace ie {
 		
 		float get_min_advance() override;
 		
-		void draw_debug(sf::RenderTarget& render_target, int indent_addition, uint hue, uint hue_offset) override;
+		void draw_debug(sf::RenderTarget& render_target, int indent_addition, size_t hue, size_t hue_offset) override;
 	
 	protected:
 		BoxPtr<IScalable> object;
