+++ lib/IE/component/IComponent/Text/BaseLine/LineInitInfo/LineInitInfo.cpp
@@ -1,6 +1,6 @@
 #include "LineInitInfo.hpp"
 
 namespace ie {
-	LineInitInfo::LineInitInfo(uint size, sf::Font& font, const sf::Color& color, sf::RenderTarget& render_target) : size(size), font(font), color(color), render_target(render_target) {
+	LineInitInfo::LineInitInfo(size_t size, sf::Font& font, const sf::Color& color, sf::RenderTarget& render_target) : size(size), font(font), color(color), render_target(render_target) {
 	}
 }
